+++ Core/Lib/PDF/PDFCore.php
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-namespace FacturaScripts\Core\Lib\Export;
+namespace FacturaScripts\Core\Lib\PDF;
 
 use Cezpdf;
 use FacturaScripts\Core\Base\DivisaTools;
